+++ backend/src/profile/application/service/patient-profile.service.ts
@@ -1,6 +1,6 @@
-import { ProtectorRegisterDto } from "src/user/interface/dto/protector-register.dto";
+import { ProtectorRegisterDto } from "src/profile/interface/dto/protector-register.dto";
 import { PatientProfileMapper } from "../mapper/patient-profile.mapper";
-import { PatientProfileRepository } from "src/user/infra/repository/patient-profile.repository";
+import { PatientProfileRepository } from "src/profile/infra/repository/patient-profile.repository";
 import { Injectable } from "@nestjs/common";
 
 @Injectable()
