+++ backend/test/unit/user/application/service/user-service.spec.ts
@@ -5,17 +5,18 @@ import { AuthService } from "src/auth/application/service/auth.service"
 import { User } from "src/user-auth-common/domain/entity/user.entity"
 import { ROLE } from "src/user-auth-common/domain/enum/user.enum"
 import { UserMapper } from "src/user/application/mapper/user.mapper"
-import { CaregiverProfileService } from "src/user/application/service/caregiver-profile.service"
-import { PatientProfileService } from "src/user/application/service/patient-profile.service"
+import { CaregiverProfileService } from "src/profile/application/service/caregiver-profile.service"
+import { PatientProfileService } from "src/profile/application/service/patient-profile.service"
 import { UserService } from "src/user/application/service/user.service"
-import { CaregiverProfileBuilder } from "src/user/domain/builder/profile.builder"
-import { CaregiverRegisterDto } from "src/user/interface/dto/caregiver-register.dto"
-import { ProtectorRegisterDto } from "src/user/interface/dto/protector-register.dto"
+import { CaregiverProfileBuilder } from "src/profile/domain/builder/profile.builder"
+import { CaregiverRegisterDto } from "src/profile/interface/dto/caregiver-register.dto"
+import { ProtectorRegisterDto } from "src/profile/interface/dto/protector-register.dto"
 import { MockAuthService } from "test/unit/__mock__/auth/service.mock"
 import { MockUserRepository } from "test/unit/__mock__/user-auth-common/repository.mock"
-import { MockCaregiverProfileService, MockPatientProfileService, MockUserMapper } from "test/unit/__mock__/user/service.mock"
+import { MockUserMapper } from "test/unit/__mock__/user/service.mock"
 import { Repository } from "typeorm"
 import { TestUser } from "../../user.fixtures"
+import { MockCaregiverProfileService, MockPatientProfileService } from "test/unit/__mock__/profile/service.mock"
 
 describe('UserService Test', () => { 
     let userService: UserService,
